case list refactor
