+++ frontend/projects/valtimo/case/src/lib/components/case-list-tabs/case-list-tabs.component.html
@@ -1,0 +1,49 @@
+<!--
+  ~ Copyright 2015-2025 Ritense BV, the Netherlands.
+  ~
+  ~ Licensed under EUPL, Version 1.2 (the "License");
+  ~ you may not use this file except in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~ https://joinup.ec.europa.eu/collection/eupl/eupl-text-eupl-12
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" basis,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  -->
+
+<ng-container *ngIf="!visibleCaseTabs; else configuredTabs">
+  <cds-tabs type="contained" class="case-list-tabs">
+    <cds-tab
+      *ngFor="let tab of defaultTabs; trackBy: trackByIndex"
+      [active]="assigneeFilter === tab"
+      heading="{{ 'case.tabs.' + tab | translate }}"
+      [id]="tab"
+      (selected)="tabChange(tab)"
+    ></cds-tab>
+  </cds-tabs>
+</ng-container>
+
+<ng-template #configuredTabs>
+  <cds-tabs type="contained" class="case-list-tabs">
+    <cds-tab
+      *ngFor="let tab of visibleCaseTabs; trackBy: trackByIndex"
+      [active]="assigneeFilter === tab"
+      heading="{{ 'case.tabs.' + tab | translate }}"
+      [id]="tab"
+      (selected)="tabChange(tab)"
+    ></cds-tab>
+  </cds-tabs>
+</ng-template>
+
+<valtimo-confirmation-modal
+  confirmButtonTextTranslationKey="case.changeTabModal.confirm"
+  contentTranslationKey="case.changeTabModal.content"
+  [outputOnConfirm]="tabChange$ | async"
+  [showModalSubject$]="showChangeTabModal$"
+  titleTranslationKey="case.changeTabModal.title"
+  (cancelEvent)="onChangeTabCancel()"
+  (confirmEvent)="onChangeTabConfirm($event)"
+></valtimo-confirmation-modal>
